feat: generate a Twitter thread from README
mplemented a new feature in the LLM_operations.R file that enables the generation of Twitter threads designed to present the repository to the community. This function reads the README.md file and uses a Large Language Model to craft a series of tweets that summarize the repository's purpose and features.

Updated the README.md and README.Rmd files to include this new feature in the list of capabilities offered by the AiGitCraft package, ensuring that users are aware of the new functionality for social media engagement.