feat: Add dependency installation check function
A new utility function, check_and_install_dependencies, has been added to the R/utils.R file. This function iterates over a list of package dependencies, checks if they are installed, and offers the user the option to install any missing packages interactively. It also handles non-interactive sessions by throwing an error if a required package is not installed.